Adjust duration unit check to include years
Improved the duration unit check functionality in the LeaderboardFunctions section. Now before checking for months, the system first checks if the count of years is over 0. This allows for a more accurate representation of longer time periods. This change will benefit scenarios where they are dealing with durations which spans for more than a year.